+++ src/legacy_puyo_tools/formats/__init__.py
@@ -1,0 +1,5 @@
+"""Formats that older Puyo Puyo games uses.
+
+SPDX-FileCopyrightText: 2025 Samuel Wu
+SPDX-License-Identifier: MIT
+"""